requirements.txt
tensorflow-gpu==2.6.0
pillow==8.3.1
scipy==1.7.1
flask==2.0.1
gunicorn==20.1.0
selenium==3.141.0
chromedriver-autoinstaller==0.2.2
numpy==1.19.5